+++ issue-tracking-orm/src/main/java/swt6/orm/logic/interfaces/EmployeeLogic.java
@@ -1,10 +1,11 @@
 package swt6.orm.logic.interfaces;
 
 import swt6.orm.domain.Employee;
+import swt6.orm.domain.Issue;
 
 public interface EmployeeLogic {
     Employee create(Employee employee);
     Employee update(Employee employee);
     void delete(Employee employee);
-    Employee findById(int id);
+    Employee findById(long id);
 }
